CLEAN: fix warnings
(cherry picked from commit 3d7b0a28558a2caeb21ae6f3cb33fac7ebfd609a)

 Conflicts:
	SofaKernel/modules/SofaBaseTopology/TetrahedronSetGeometryAlgorithms.inl
	modules/SofaGeneralEngine/GenerateGrid.inl
	modules/SofaGeneralEngine/SmoothMeshEngine.inl
	modules/SofaGeneralTopology/BezierTetrahedronSetGeometryAlgorithms.inl
	modules/SofaMiscMapping/CenterOfMassMapping.inl
	modules/SofaMiscMapping/DistanceFromTargetMapping.h
	modules/SofaMiscMapping/DistanceFromTargetMapping.inl

+++ SofaKernel/modules/SofaBaseMechanics/SubsetMapping.inl
@@ -223,7 +223,7 @@ void SubsetMapping<TIn, TOut>::apply ( const core::MechanicalParams* /*mparams*/
 {
     const IndexArray& indices = f_indices.getValue();
     
-    if (f_resizeToModel.getValue() || this->toModel->getSize() < indices.size())
+    if (f_resizeToModel.getValue() || this->toModel->getSize() < (int)indices.size())
     { 
         if (this->toModel->getSize() != indices.size()) 
         {
